Event_Registraion_system/event_project/requirements.txt
asgiref==3.8.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
